add unicorn configuration for production environment
This uses environment variables to control configuration values
such as the path the the unicorn socket and the directory where the
app is deployed. This allows the ansible playbook (which will build
and manage the production server) to ensure those values are set
consistently across various files (e.g., the unicorn systemd service
configuration, the nginx configuration).