change git_source in Gemfile to use HTTPS instead of SSH
The use of SSH was breaking some automated tests I have been running.
For example, on a fresh VM the `bundle install` step cannot be run
non-interactively due to the host key checking stage when using SSH.

This will resolve those issues and ensure `bundle install` can be
run non-interactively for testing purposes.

As far as I can tell there is no downside to using https for gems,
and when I asked Vlad he saw no issue with it either.