add capistrano production deployment step to CircleCI
This will only be run on the branch specified by the environment
variable BRANCH_TO_DEPLOY_TO_PRODUCTION (in CircleCI). This is
used to test the capistrano deployment.

We will most likely not use CircleCI for performing deployments like
this to our production environment because we cannot do that without
making some changes to our security configuration that we do not want
to make (opening up SSH access to a wide range of IPs, or using a
jump host). See here for more: https://support.circleci.com/hc/en-us/articles/115014372807-IP-Address-ranges-for-whitelisting-

This is however very useful for testing the capistrano deployment
with a staging server I have temporarily set up.